Expo-AV And AppState
For expo-av, we dont have to worry about releasing resources when our app goes from active to background. It is already handles for us out of the box.
Eg: when we play a sound and leave the app, the sound will immidiately stop.
- On our own we could configure Audio.setAudioModeAsync() to allow our sound to play even if the phone is in background

But in real projects, our usecases of appStates are:
- You probably only want to show an animation on the screen if the user is actively looking at your activity and you want to stop that animation as soon as the screen is not visible anymore
- Another Example is like in Gmail app, they save your draft email, and when you come back to the screen, it like you never left that draft page
- Understanding AppStates gives you as the developer an opportunity to intervain and execute statements. Things like restoring or saving any data or even cleaning up references that you no longer need at the moments that you ned them